docs - clarify difference between "replace" and "update" actions
### Why this PR?

Developers might ask: "what is the difference between replace and update?". This PR seeks to make that answer explicitly clear. Otherwise devs would be forced to open a new tab to find the relevant line in the source code - which is a little bit of a pain and not the ideal time (when one is reading the handbook (IMO)). 

PR based on the discussion [here](https://discuss.hotwired.dev/t/difference-between-replace-and-update-turbo-streams/3148/2). 

My thanks to contributors to the library and disucssion boards.

If this PR is not useful / or incorrect etc. feel free to close without compunction.

rdgs

Ben