Document the method for terminating a form submission
This supports the pattern where form values are validated in javascript, and the form submission is aborted if validation fails. It works with the current code, but is not documented. I'm hoping that by making it part of the documented API, it can be relied upon not to change in future versions.