Mention Rails in more places
[This tweet](https://twitter.com/bradgessler/status/1630655623192002560) is good feedback, that if you're getting started it could be more clear how Turbo works with Rails. This PR makes some minor tweaks to the site to mention Rails in a few more places and give it more weight. While you _can_ use Turbo without Rails or with other frameworks, the reality is the overwhelming majority of users will use it with Rails.

If we want to go deeper on this idea, a page called "Using Turbo with Rails" might be a good next step. It would re-state a lot of https://github.com/hotwired/turbo-rails but I think it would not be a bad thing to have all those docs in one place. Happy to prepare another PR for this if there's appetite.