Document Type information for Custom Events
It appears that the team's [official stance][] is that documentation for
the package's Custom Events should reside on a consumer-facing website.

The documentation for Custom Events already exists, but doesn't provide
particulars for the shapes of the most of the `event.details` objects.

This commit brings more structure to the page, namely:

* groups events by their source
* promotes event names to deeply-linkable headings
* adds HTML tables to describe each event's `event.detail` properties,
  along with type information

[official stance]: https://github.com/hotwired/turbo/pull/984#issuecomment-1711491766