Rename "InstantClick" section
Turbo's link prefetching implementation is in no way affiliated with the
pre-existing [InstantClick][] package. Turbo doesn't have a client-side
dependency on the [instantclick package][], nor does it share any code
with the [dieulot/instantclick][] repository.

With this in mind, this commit removes all mentions of "InstantClick"
from the site and renames the heading to "Prefetching Links on Hover".

While editing that section, this commit also removes mention of handling
`touchstart` events, since those events are not accounted for in the
current implementation.

[InstantClick]: http://instantclick.io
[instantclick package]: https://www.npmjs.com/package/instantclick
[dieulot/instantclick]: https://github.com/dieulot/instantclick